Permission to delete commnents

+++ comments/views.py
@@ -1,17 +1,33 @@
 from django.shortcuts import render, get_object_or_404
 from django.contrib.contenttypes.models import ContentType
-from django.http import HttpResponseRedirect
+from django.http import HttpResponseRedirect, Http404, HttpResponse
 from .forms import CommentForm
 from .models import Comment
 from django.contrib import  messages
 # Create your views here.
 def comment_delete(request, id):
-    obj = get_object_or_404(Comment, id=id)
+    #obj = get_object_or_404(Comment, id=id)
+    #obj = Comment.objects.get(id=id)
+
+    try:
+        obj = Comment.objects.get(id=id)
+    except:
+        raise Http404
+
+    if obj.user != request.user:
+        #messages.success(request, "You do not have permission to view this.")
+        #raise Http404
+        response = HttpResponse("You do not have permission to do this.")
+        response.status_code = 403
+        return response
+
+
     if request.method == "POST":
         parent_obj_url = obj.content_object.get_absolute_url()
         obj.delete()
         messages.success(request, "This has been deleted.")
         return HttpResponseRedirect(parent_obj_url)
+
     context = {
         "object": obj
     }
@@ -20,7 +36,17 @@ def comment_delete(request, id):
 
 
 def comment_thread(request, id):
-    obj = get_object_or_404(Comment, id = id)
+    #obj = Comment.objects.get(id=id)
+
+    try:
+        obj = Comment.objects.get(id=id)
+    except:
+        raise Http404
+
+    if not obj.is_parent:
+        obj = obj.parent
+
+
     content_object = obj.content_object
     content_id = obj.content_object.id
     initial_data = {
